Add DI, EF Core, Swagger, and refactor services
- Introduced Dependency Injection and configured EF Core with SQL Server.
- Added Swagger for API documentation and updated `Program.cs`.
- Refactored `UserService` and `IUserService` to use DTOs.
- Implemented `UserController` for user-related API endpoints.
- Added `TodoItemsController` and `AppDbContext` for `TodoItem` management.
- Created initial database migration and model snapshot.
- Updated project files with necessary NuGet packages.
- Removed unused `WeatherForecastController` and `DTOs` folder.
- Updated `appsettings.json` with connection strings.
- Changed `Product.Price` type from `decimal` to `double`.

+++ SSC.EShop/SSC.EShop.Core/Entities/Product.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,15 +8,15 @@ namespace SSC.EShop.Core.Entities
 {
     public class Product
     {
-        
+
         public Guid Id { get; set; }
-      
+
         public string Name { get; set; }
-        
+
         public string Description { get; set; }
-      
-        
-        public decimal Price { get; set; }
+
+
+        public double Price { get; set; }
